Refactor Contract model
Move some non logic code from the core of the Contract model
into a number of concerns.
